bug855: New test methods and data for EDConfiguration
