+++ kernel/tests/data/EDConfiguration/XSConfiguration_TestStaticConfiguration.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" ?>
+<!-- EDPluginFactory test XML file -->
+<XSConfiguration> 
+  <XSImportConfiguration>
+  	<name>XSConfiguration_TestImport1</name>
+  </XSImportConfiguration>	
+  <XSPluginList>		
+    <XSPluginItem> 			
+      <name>EDPluginTestPluginFactory</name>  			
+      <XSParamList>				
+        <XSParamItem>					
+          <name>testItemName</name>					
+          <value>anotherTestValue</value>				
+        </XSParamItem>			
+      </XSParamList>		
+    </XSPluginItem>	
+    <XSPluginItem> 			
+      <name>EDPluginTestStaticConfiguration</name>  			
+      <XSParamList>				
+        <XSParamItem>					
+          <name>testItemName</name>					
+          <value>anotherTestValue</value>				
+        </XSParamItem>			
+      </XSParamList>		
+    </XSPluginItem>	
+  </XSPluginList>
+  </XSPluginList>
+</XSConfiguration>
